intraday_trend/requirements.txt
black==21.12b0
matplotlib==3.5.1
numpy==1.22.0
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
yml==0.0.1
